refactor(auth): 统一登录注册日志消息字段及时间类型为浮点数
- 登录注册日志中的 fail_reason 字段改为 message，优化日志信息统一性
- 修改后端各模型中 created_at、updated_at、last_login_time 等字段类型为 float，保持时间精度一致
- 更新各模型及对应的 Schema 中时间字段类型，及相关仓库更新操作中时间赋值
- 保持登录成功和注册成功日志消息的统一格式，增强日志可读性
- 修正日志服务中参数名称及返回结果处理，提升代码一致性

+++ backend/app/repositories/role_repository.py
@@ -41,7 +41,7 @@ class RoleRepository:
         role = await Role.get(role_id)
         if not role:
             return None
-        update_data["updated_at"] = int(time.time() * 1000)
+        update_data["updated_at"] = time.time() * 1000
         await role.update({"$set": update_data})
         return await Role.get(role_id)
